+++ R/LaTeX.R
@@ -44,8 +44,7 @@ to_latex.huxtable <- function (ht, tabular_only = FALSE, ...){
   }
 
   cap <- if (! is.na(cap <- caption(ht))) {
-    hpos <- sub('.*(left|center|right)', '\\1', caption_pos(ht))
-    if (! hpos %in% c('left', 'center', 'right')) hpos <- position(ht)
+    hpos <- get_caption_hpos(ht)
     cap_setup <- switch(hpos,
             left   = 'raggedright',
             center = 'centering',
@@ -300,7 +299,7 @@ build_clines_for_row <- function(ht, row, collapsed_borders, cb_colors) {
       blank_line_color[dc$display_col:dc$end_col] <- background_color(ht)[dc$display_row, dc$display_col]
   }
 
-  widths <- collapsed_borders$horiz[row + 1,]
+  widths <- collapsed_borders$horiz[row + 1, ]
   if (all(widths == 0)) {
     return('')
   } else {
@@ -348,6 +347,5 @@ v_border <- function (ht, row, col, collapsed_borders, cb_colors) {
   color <- cb_colors$vert[row, col]
   color <- format_color(color, default = 'black')
 
-  paste0('!{\\color[RGB]{', color ,'}\\vrule width ', width, 'pt}')
+  paste0('!{\\color[RGB]{', color, '}\\vrule width ', width, 'pt}')
 }
-