+++ R/attributes.R
@@ -102,7 +102,7 @@ make_getter_setters <- function(attr_name, attr_type = c('cell', 'row', 'col', '
             if (nargs == 3) {
               if (missing(value)) value <- col
               if (! is.matrix(row)) stop('No columns specified, but `row` argument did not evaluate to a matrix')
-              if (byrow) stop('byrow = TRUE makes no sense if `row` is a matrix')
+              if (byrow) stop('`byrow = TRUE` makes no sense if `row` is a matrix')
               .(attr_symbol)(ht)[row] <- value
             } else {
               if (nargs == 2) {
@@ -218,9 +218,9 @@ get_default_properties <- function (names = NULL) {
 #'
 #' @examples
 #' ht <- hux(a = 1:3, b = 1:3)
-#' ht <- set_cell_properties(ht, 1, 1, font = 'Palatino', font_size = 14)
-#' font(ht)
-#' font_size(ht)
+#' ht <- set_cell_properties(ht, 1, 1, italic = TRUE, text_color = 'red')
+#' text_color(ht)
+#' ht
 set_cell_properties <- function (ht, row, col, ...) {
   props <- list(...)
   if (! all(names(props) %in% huxtable_cell_attrs)) stop('Unrecognized properties: ', paste(setdiff(names(props),
@@ -260,7 +260,7 @@ make_getter_setters('valign', 'cell', check_fun = is.character, check_values = c
 #' @templateVar value_param_desc A character vector or matrix which may be 'left', 'center', 'right' or \code{NA}.
 #' @template getset-example
 #' @templateVar attr_val 'right'
-#' @template getset-rowspec-example
+#' @template getset-visible-rowspec-example
 #' @templateVar attr_val2 'left'
 #' @export align align<- set_align align.huxtable align<-.huxtable
 #' @S3method align huxtable
@@ -357,10 +357,10 @@ check_span_shadows <- function (ht, rc, value) {
 #' @template getset-cell
 #' @templateVar attr_name background_color
 #' @templateVar attr_desc Background color
-#' @templateVar value_param_desc A vector or matrix of R colors.
+#' @templateVar value_param_desc A character vector or matrix of valid R color names.
 #' @template getset-example
 #' @templateVar attr_val grey(.95)
-#' @template getset-rowspec-example
+#' @template getset-visible-rowspec-example
 #' @templateVar attr_val2 'yellow'
 #' @family formatting functions
 #' @export background_color background_color<- set_background_color background_color.huxtable background_color<-.huxtable
@@ -372,10 +372,10 @@ make_getter_setters('background_color', 'cell')
 #' @template getset-cell
 #' @templateVar attr_name text_color
 #' @templateVar attr_desc Text color
-#' @templateVar value_param_desc A vector or matrix of R colors.
+#' @templateVar value_param_desc A character vector or matrix of valid R color names.
 #' @template getset-example
-#' @templateVar attr_val 'navy'
-#' @template getset-rowspec-example
+#' @templateVar attr_val 'blue'
+#' @template getset-visible-rowspec-example
 #' @templateVar attr_val2 'red'
 #' @family formatting functions
 #' @export text_color text_color<- set_text_color text_color.huxtable text_color<-.huxtable
@@ -397,7 +397,7 @@ make_getter_setters('text_color', 'cell')
 #' @template getset-example
 #' @templateVar attr_val 1
 #' @templateVar extra print_screen(ht)
-#' @template getset-rowspec-example
+#' @template getset-visible-rowspec-example
 #' @templateVar attr_val2 2
 #' @export left_border left_border<- set_left_border left_border.huxtable left_border<-.huxtable
 #' @S3method left_border huxtable
@@ -457,7 +457,7 @@ make_getter_setters('bottom_border', 'cell', check_fun = is.numeric)
 #' @seealso \code{\link{left_border}}, \code{\link{set_outer_borders}}
 #' @examples
 #' ht <- huxtable(a = 1:3, b = 1:3)
-#' ht <- set_all_borders(ht, 1:3, 1:2, 1)
+#' set_all_borders(ht, 1:3, 1:2, 1)
 set_all_borders <- function(ht, row, col, value, byrow = FALSE) {
   call <- sys.call()
   for (set_b in paste0('set_', c('top', 'bottom', 'left', 'right'), '_border')) {
@@ -537,7 +537,7 @@ get_all_borders <- function(ht, row, col) {
 #' adjoining border of the previous cell to width 0 (e.g. for a left border color, unset the right border
 #' of the cell on the left).
 #' @seealso \code{\link{set_all_border_colors}}
-#' @template getset-rowspec-example
+#' @template getset-visible-rowspec-example
 #' @templateVar attr_val2 'blue'
 #' @export left_border_color left_border_color<- set_left_border_color left_border_color.huxtable left_border_color<-.huxtable
 #' @S3method left_border_color huxtable
@@ -769,7 +769,7 @@ make_getter_setters('na_string', 'cell', check_fun = is.character)
 #' @template getset-example
 #' @templateVar attr_val TRUE
 #' @templateVar extra print_screen(ht)
-#' @template getset-rowspec-example
+#' @template getset-visible-rowspec-example
 #' @templateVar attr_val2 FALSE
 #' @family formatting functions
 #' @export bold bold<- set_bold bold.huxtable bold<-.huxtable
@@ -868,7 +868,7 @@ make_getter_setters('rotation', 'cell', check_fun = is.numeric)
 #' ht_bands # probably not what you want
 #' number_format(ht_bands) <- NA
 #' ht_bands
-#' @template getset-rowspec-example
+#' @template getset-visible-rowspec-example
 #' @templateVar attr_val 2
 #' @templateVar attr_val2 3
 #' @export number_format number_format<- set_number_format number_format.huxtable number_format<-.huxtable
@@ -880,7 +880,6 @@ make_getter_setters('number_format', 'cell')
 # override the default
 `number_format<-.huxtable` <- function(ht, value) {
   stopifnot(all(sapply(value, function (x) is.numeric(x) || is.character(x) || is.function(x) || is.na(x) )))
-  # if (is.atomic(value) || is.list(value)) value[is.na(value)] <- huxtable_env$huxtable_default_attrs[['number_format']]
   attr(ht, 'number_format')[] <- value
   ht
 }
@@ -1045,6 +1044,9 @@ make_getter_setters('tabular_environment', 'table', check_fun = is.character)
 #' A length-one character vector to be used as a table label in LaTeX, or as an ID for the table in HTML. Set to \code{NA} to remove any label.
 #' @template getset-example
 #' @templateVar attr_val 'tab:mytable'
+#' @details
+#' LaTeX table labels typically start with "tab:", and they must do so if you want table numbering
+#' in \href{http://bookdown.org}{bookdown}.
 #' @export label label<- set_label label.huxtable label<-.huxtable
 #' @S3method label huxtable
 #' @S3method label<- huxtable